components refactor + some other config
