+++ package.json
@@ -47,7 +47,8 @@
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.16",
     "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11"
+    "@types/react-dom": "^18.0.11",
+    "eslint": "^8.41.0"
   },
   "scripts": {
     "dev": "next dev",
