--- src/components/Panels/index.ts
@@ -1,1 +1,0 @@
-export * from './OutcomePanel';
