+++ src/components/Prompts/index.ts
@@ -1,0 +1,6 @@
+// Barrel File
+export * from './ButtonGroup/ButtonGroup';
+export * from './CurrentPrompt/CurrentPrompt';
+export * from './MultiSelect/MultiSelect';
+export * from './PreviousAnswers/PreviousAnswers';
+export * from './PromptPanel/PromptPanel';
