+++ src/components/ui/RichTextOutput/RichTextOutput.tsx
@@ -1,5 +1,5 @@
 import { TypographyStylesProvider } from '@mantine/core';
-import { JSONContent, generateHTML, Extensions } from '@tiptap/core';
+import { Extensions, JSONContent, generateHTML } from '@tiptap/core';
 import { Blockquote } from '@tiptap/extension-blockquote';
 import { Bold } from '@tiptap/extension-bold';
 import { BulletList } from '@tiptap/extension-bullet-list';
@@ -13,11 +13,11 @@ import { Heading } from '@tiptap/extension-heading';
 import { History } from '@tiptap/extension-history';
 import { HorizontalRule } from '@tiptap/extension-horizontal-rule';
 import { Italic } from '@tiptap/extension-italic';
+import { Link } from '@tiptap/extension-link';
 import { ListItem } from '@tiptap/extension-list-item';
 import { OrderedList } from '@tiptap/extension-ordered-list';
 import { Paragraph } from '@tiptap/extension-paragraph';
 import { Strike } from '@tiptap/extension-strike';
-import { Link } from '@tiptap/extension-link';
 import { Text } from '@tiptap/extension-text';
 import { Underline } from '@tiptap/extension-underline';
 import { FC } from 'react';
